+++ lib/modules/settings/routes/configuration_lidarr/route.dart
@@ -36,14 +36,6 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     return LunaAppBar(
       scrollControllers: [scrollController],
       title: 'Lidarr',
-      actions: [
-        LunaIconButton(
-          icon: Icons.help_outline_rounded,
-          onPressed: () async {
-            SettingsDialogs().moduleInformation(context, LunaModule.LIDARR);
-          },
-        ),
-      ],
     );
   }
 
@@ -51,6 +43,7 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     return LunaListView(
       controller: scrollController,
       children: [
+        LunaModule.LIDARR.informationBanner(),
         _enabledToggle(),
         _connectionDetailsPage(),
         const LunaDivider(),