+++ lib/modules/settings/routes/configuration_tautulli/route.dart
@@ -36,13 +36,6 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     return LunaAppBar(
       title: 'Tautulli',
       scrollControllers: [scrollController],
-      actions: [
-        LunaIconButton(
-          icon: Icons.help_outline_rounded,
-          onPressed: () async =>
-              SettingsDialogs().moduleInformation(context, LunaModule.TAUTULLI),
-        ),
-      ],
     );
   }
 
@@ -50,6 +43,7 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     return LunaListView(
       controller: scrollController,
       children: [
+        LunaModule.TAUTULLI.informationBanner(),
         _enabledToggle(),
         _connectionDetailsPage(),
         const LunaDivider(),